requirements.txt
anyio==3.6.1
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
click==8.1.3
defusedxml==0.7.1
executing==0.8.3
fastapi==0.81.0
h11==0.13.0
idna==3.3
importlib-resources==5.4.0
ipython-genutils==0.2.0
json5==0.9.6
jupyterlab-pygments==0.1.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbformat==5.1.3
notebook-shim==0.1.0
pandocfilters==1.5.0
pickleshare==0.7.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.0
python-dotenv==0.20.0
pyzmq==22.3.0
Send2Trash==1.8.0
sniffio==1.2.0
stack-data==0.2.0
starlette==0.19.1
tornado==6.1
typing_extensions==4.3.0
uvicorn==0.18.3
wcwidth==0.2.5
webencodings==0.5.1
